refactor: Add Playwright installation step to Dockerfile
This commit adds a new step to the Dockerfile to install Playwright. Playwright is a powerful automation library that allows for browser automation and testing. By including this step in the Dockerfile, we ensure that Playwright is installed and ready to be used in the application. This change enhances the functionality of the application by enabling the use of Playwright for automated testing and crawling.